services/ai/pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = "AI (Artificial Intelligence) service for TheraVillage"
authors = ["Daniel Nurieli <daniel_nurieli@alumni.brown.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
openai = "^1.3.7"
google-cloud-secret-manager = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
